fix(product): widen st_mode before masking so the macOS legs compile
`mode_t` is `u16` on Darwin and `u32` on Linux, so `stat.st_mode & 0o777` typed
against a `u32` field compiles on one platform and not the other. Both macOS
release legs failed to build on it:

    error[E0308]: mismatched types
       --> crates/vaultspec-product/src/generation/authority.rs:155:19
    155 |             mode: stat.st_mode & 0o777,
        |                   ^^^^^^^^^^^^^^^^^^^^ expected `u32`, found `u16`

`u32::from` is a widening conversion on Darwin and the identity on Linux, so one
expression is correct on both. It is deliberately `from` rather than `as`: a cast
would keep compiling and silently truncate if the platform's width ever changed,
which is the failure this line already demonstrates is easy to miss.

The masking is unchanged - `& 0o777` still yields the permission bits.

+++ engine/crates/vaultspec-product/src/generation/authority.rs
@@ -152,7 +152,7 @@ impl DirectoryAuthority {
             is_directory: rustix::fs::FileType::from_raw_mode(stat.st_mode)
                 == rustix::fs::FileType::Directory,
             owner: stat.st_uid,
-            mode: stat.st_mode & 0o777,
+            mode: u32::from(stat.st_mode) & 0o777,
         })
     }
 
@@ -182,7 +182,7 @@ impl DirectoryAuthority {
             is_directory: rustix::fs::FileType::from_raw_mode(stat.st_mode)
                 == rustix::fs::FileType::Directory,
             owner: stat.st_uid,
-            mode: stat.st_mode & 0o777,
+            mode: u32::from(stat.st_mode) & 0o777,
         })
     }
